add squashfs support current status

+++ uboot/squashfs/sqfs_decompressor.c
@@ -30,15 +30,10 @@ static void zlib_decompression_status(int ret)
 	}
 }
 
-int sqfs_decompress(int comp_type, void *dest, unsigned long int *dest_len,
-		    void *source, uint32_t lenp)
+int sqfs_decompress(u16 comp_type, void *dest, unsigned long *dest_len,
+		    void *source, u32 lenp)
 {
 	int ret = 0;
-//	printf("SQFS_DECOMPRESS\n");
-//	printf("dest: %p\n", dest);
-//	printf("dest_len: %ld\n", *dest_len);
-//	printf("source: %p\n", source);
-//	printf("source_len: %d\n\n", lenp);
 
 	switch (comp_type) {
 	case SQFS_COMP_ZLIB:
